Restructure table directory
