+++ packages/design-system/src/components/table/view/columnMenu/columnList.tsx
@@ -18,14 +18,13 @@
  * External dependencies.
  */
 import React, { useEffect } from 'react';
-import { PreferenceDataValues } from '@ps-analysis-tool/common';
+import { PreferenceDataValues, noop } from '@ps-analysis-tool/common';
 
 /**
  * Internal dependencies.
  */
 import ColumnListItem from './columnListItem';
-import type { TableOutput } from '../useTable';
-import { noop } from '../../../utils';
+import { TableOutput } from '../../hooks';
 
 interface ColumnListProps {
   table: TableOutput;
