+++ packages/design-system/src/components/table/view/columnMenu/columnListItem.tsx
@@ -19,11 +19,11 @@
  */
 import React from 'react';
 import classNames from 'classnames';
+import { TableColumn } from '../../hooks';
 
 /**
  * Internal dependencies.
  */
-import type { TableColumn } from '../useTable';
 
 interface ColumnListItemProps {
   column: TableColumn;
