+++ packages/design-system/src/components/table/view/columnMenu/index.tsx
@@ -19,15 +19,14 @@
  */
 import React, { useEffect, useRef, useState } from 'react';
 import classNames from 'classnames';
-import { type PreferenceDataValues } from '@ps-analysis-tool/common';
+import { noop, type PreferenceDataValues } from '@ps-analysis-tool/common';
 import { createPortal } from 'react-dom';
 
 /**
  * Internal dependencies.
  */
 import ColumnList from './columnList';
-import type { TableOutput } from '../useTable';
-import { noop } from '../../../utils';
+import { TableOutput } from '../../hooks';
 
 interface ColumnMenuProps {
   table: TableOutput;
