+++ packages/design-system/src/components/table/view/index.tsx
@@ -25,8 +25,8 @@ import { PreferenceDataValues } from '@ps-analysis-tool/common';
 import TableHeader from './tableHeader';
 import TableBody from './tableBody';
 import ColumnMenu from './columnMenu';
-import { TableData, TableOutput, TableRow } from './useTable';
-import { noop } from '../../utils';
+import { TableData, TableOutput, TableRow } from '../hooks';
+import { noop } from '../../../utils';
 
 interface TableProps {
   table: TableOutput;
