+++ packages/design-system/src/components/table/view/tableBody/bodyCell.tsx
@@ -22,7 +22,7 @@ import React from 'react';
 /**
  * Internal dependencies.
  */
-import type { InfoType } from '../useTable';
+import { InfoType } from '../../hooks';
 interface BodyCellProps {
   cell: React.JSX.Element | InfoType;
   width: number;
