+++ packages/design-system/src/components/table/view/tableBody/bodyRow.tsx
@@ -25,7 +25,7 @@ import { CookieTableData } from '@ps-analysis-tool/common';
  * Internal dependencies.
  */
 import BodyCell from './bodyCell';
-import type { TableColumn, TableRow } from '../useTable';
+import { TableColumn, TableRow } from '../../hooks';
 
 interface BodyRowProps {
   row: TableRow;
