+++ packages/design-system/src/components/table/view/tableHeader/headerResizer.tsx
