+++ packages/design-system/src/components/table/view/tableHeader/index.tsx
@@ -18,13 +18,12 @@
  * External dependencies.
  */
 import React, { useCallback } from 'react';
-import { PreferenceDataValues } from '@ps-analysis-tool/common';
+import { PreferenceDataValues, noop } from '@ps-analysis-tool/common';
 /**
  * Internal dependencies.
  */
 import HeaderRow from './headerRow';
-import type { TableOutput } from '../useTable';
-import { noop } from '../../../utils';
+import { TableOutput } from '../../hooks';
 
 interface TableHeaderProps {
   table: TableOutput;
