+++ packages/extension/src/view/devtools/components/cookies/cookiesListing/index.tsx
@@ -23,7 +23,6 @@ import { Resizable } from 're-resizable';
  * Internal dependencies.
  */
 import { useCookieStore } from '../../../stateProviders/syncCookieStore';
-
 import { useFilterManagementStore } from '../../../stateProviders/filterManagementStore';
 import ChipsBar from '../cookieFilter/chips';
 import CookieTopBar from '../cookieTopBar';
